extracted cell highlighting from hover on plot circles to map.
