+++ src/map.js
@@ -19,3 +19,34 @@ export function renderHeatLayer(siteAggs, heatLayerGroup, map) {
   map.addLayer(heatLayerGroup);
   return heatLayer;
 }
+
+export function highlightLayer(layer) {
+  layer.setStyle({
+    weight: 5,
+    opacity: 0.9
+  });
+  if (!L.Browser.ie && !L.Browser.opera && !L.Browser.edge) {
+    layer.bringToFront();
+  }
+}
+
+/**
+ * centralised place to store value
+ */
+function getOutlineOpacity() {
+  return 0.15;
+}
+
+/**
+ * Resets layer outline weight and opacity to original values.
+ * Values are hardcoded due to geojson.reset() not working as planned.
+ * @param {*} layer
+ */
+export function disableHighlightLayer(layer) {
+  var properties = layer.feature.properties;
+  //console.log(properties);
+  layer.setStyle({
+    weight: 1,
+    opacity: getOutlineOpacity(properties.hasSamples)
+  });
+}
